% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams2olat.R
\name{exams2olat}
\alias{exams2olat}
\title{Prepare exercises to be uploaded to OLAT}
\usage{
exams2olat(file, n = 1L, dir = ".", name = "olattest",
  maxattempts = 1L, cutvalue = 1000, solutionswitch = FALSE,
  stitle = "Aufgabe", ititle = "Frage",
  adescription = "Bitte bearbeiten Sie folgende Aufgaben.",
  sdescription = "Bitte beantworten Sie folgende Frage.",
  eval = list(partial = FALSE, negative = FALSE), ...)
}
\arguments{
\item{file}{list, of \code{md}/\code{Rmd} files to be used}

\item{n}{integer, number of randomized tests to be created (default \code{1L})}

\item{dir}{character, where to store the resulting file(s) (default \code{.})}

\item{name}{character name of the test/quiz}

\item{maxattempts}{integer, he maximum attempts for one question (must be
smaller than \code{100000L}).}

\item{cutvalue}{numeric, the cutvalue at which the exam is passed}

\item{solutionswitch}{logical Should the question/item solutionswitch be
enabled? In OLAT this means that the correct solution is
shown after an incorrect solution was entered by an examinee
(i.e., this is typically only useful if \code{maxattempts = 1}).}

\item{stitle}{character A title that should be used for the sections. May
be a vector of length 1 to use the same title for each
section, or a vector containing different section titles.}

\item{ititle}{character A title that should be used for the assessment
items. May be a vector of length 1 to use the same title for
each item, or a vector containing different item titles. Note
that the maximum of different item titles is the number of
sections/questions that are used for the exam.}

\item{adescription}{character  Description (of length 1) for the overall
assessment (i.e., exam).}

\item{sdescription}{character Vector of descriptions for each section.}

\item{eval}{named list, specifies the settings for the evaluation policy,
see function \code{\link[exams]{exams_eval}}}
}
\description{
Takes (a set of) exams questions (md/Rmd) files and creates
\code{n} randomized tests to be used in Olat.
Outputs an \code{rds} and \code{zip} file. The \code{rds} file
contains the questions/answers of the randomized tests and is
REQUIRED for evaluation (keep it; use a seed to make it
reproduceable), the zip archive file contains the files which
for the OLAT import.
}
