% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams2openolat.R
\name{exams2openolat}
\alias{exams2openolat}
\title{Generation of Exams for OpenOlat (Uni Innsbruck)}
\usage{
exams2openolat(
  file,
  n = 1L,
  dir = ".",
  name = "olattest",
  maxattempts = 1,
  cutvalue = 1000,
  solutionswitch = FALSE,
  qti = "2.1",
  stitle = "Aufgabe",
  ititle = "Frage",
  adescription = "",
  sdescription = "",
  eval = list(partial = FALSE, negative = FALSE),
  template = "qti21",
  ...
)
}
\arguments{
\item{file}{character. A specification of a (list of) exercise files.}

\item{n}{integer. The number of copies to be compiled from \code{file}.}

\item{dir}{character. The default is either display on the screen or the
current working directory.}

\item{name}{character. A name prefix for resulting ZIP and RDS file.}

\item{maxattempts}{integer. The maximum attempts for one question, may also
be set to \code{Inf}.}

\item{cutvalue}{numeric. The cutvalue at which the exam is passed.}

\item{solutionswitch}{logical. Should the question/item solutionswitch be
enabled?}

\item{qti}{character indicating whether QTI \code{"1.2"} or \code{"2.1"}
(default) should be generated.}

\item{stitle, ititle, adescription, sdescription}{character. Descriptions
for various titles/descriptions. Defaults to generic German titles.}

\item{eval}{named list. Specifies the settings for the evaluation policy, see function
\code{\link{exams_eval}}. The default is not to partial credits for multiple-choice
exercises.}

\item{template}{character. The IMS QTI 2.1 template that should be used. In addition
to the default template this can be set to \code{"qti21test"} which uses almost
the same template but with a PASS variable. Thus, no \code{cutvalue} is needed
and the online tests just generate a SCORE.}

\item{\dots}{arguments passed on to \code{\link[exams]{exams2openolat}}.}
}
\value{
A list of exams as generated by \code{\link[exams]{xexams}} is
returned invisibly.
}
\description{
Interfaces \code{\link[exams]{exams2openolat}} with
slightly different default values as used at the Department of
Statistics, Uni Innsbruck.
}
\details{
\code{exams2openolat} is a convenience interface for
\code{\link[exams]{exams2openolat}} with somewhat different defaults:
German titles/descriptions, partial credits disabled, solution switch turned off,
large cut value (so that the test cannot be passed), and with fancy
quotes turned off in verbatim R output. Finally, an RDS file is stored
as a by-product containing the \code{\link[exams]{xexams}} list. This
enables extracting and displaying specific exercises from an online test in R.
}
\examples{

## load package and enforce par(ask = FALSE)
library("exams")
options(device.ask.default = FALSE)

## define an exams (= list of exercises)
myexam <- list(
  "boxplots",
  c("tstat", "ttest", "confint"),
  c("regression", "anova"),
  c("scatterplot", "boxhist"),
  "relfreq"
)

## output directory
mydir <- tempdir()

## generate .zip with German OpenOlat exam in temporary directory
## using a few customization options
exams2openolat(myexam, n = 3, dir = mydir, maxattempts = 2)
dir(mydir)

}
\seealso{
olat_eval olat_exercise
}
\keyword{utilities}
