% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olat_feedback.R
\name{nops_feedback}
\alias{nops_feedback}
\title{Generate nops test feedback files (answer/solution)}
\usage{
nops_feedback(res, xexam, name = "nops_feedback")
}
\arguments{
\item{res}{\code{data.frame}, content of the csv file created
by \code{\link[exams]{nops_eval}}.}

\item{xexam}{list as returned from reading the rds file}

\item{name}{character, name of the test, will be used to name
the zip archive file and the html files}
}
\value{
Returns the name of the zip file created.
}
\description{
Similar to what \code{\link[c403]{olat_eval}} does but in a more
detailed way. Creates html files for each participant with feedback
about his/her test results (including questions and solutions).
}
\author{
Reto Stauffer
}
