% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nops_itemresp.R
\name{nops_itemresp}
\alias{nops_itemresp}
\title{Extract Item-Response Data from NOPS Exams}
\usage{
nops_itemresp(
  eval = "nops_eval.csv",
  exam = Sys.glob("*.rds"),
  psychotools = NULL,
  labels = NULL,
  ...
)
}
\arguments{
\item{eval}{character. File name of CSV output from \code{nops_eval}.}

\item{exam}{character. File name of RDS output from \code{exams2nops}.}

\item{psychotools}{logical. Should \code{itemresp} from \pkg{psychotools} be
used as the class for item response columns? By default, this is used if the
\pkg{psychotools} package is available. If \code{FALSE}, matrices with dummy
codings are used.}

\item{labels}{function for extracting exercise labels from each
\code{$metainfo}.  By default the original file name \code{$metainfo$file}
is used.}

\item{\dots}{additional arguments (such as \code{colClasses}) passed to
\code{read.csv2(eval, dec = ".", ...)}.}
}
\value{
A \code{data.frame}.
}
\description{
Process data from NOPS evaluation results (via \code{\link[c403]{nops_eval}})
for subsequent IRT (item response theory) modeling.
}
\details{
\code{nops_itemresp} returns a data frame with several item response
outcomes for each student: \code{solved} indicates whether or not an
exercise was fully solved, \code{partial} whether or not it was at least
partially solved.  \code{points} gives the points achieved for each
exercise. The corresponding \code{nsolved}, \code{npartial}, and
\code{npoints} are the sums of these for each student. Moreover,
\code{solved2}, \code{partial2}, and \code{points2} distinguish not only the
exercises within the exam but also the actual source template within each
exercise.
}
\seealso{
nops_eval
}
\keyword{utilities}
