% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olat_eval.R
\name{olat_eval}
\alias{olat_eval}
\title{Evaluate OLAT Exams}
\usage{
olat_eval(file, plot = TRUE, export = FALSE)
}
\arguments{
\item{file}{character. Base file name of RDS and XLS file with
exam generated by \code{\link[c403]{exams2olat}} and exported
OLAT results, respectively.}

\item{plot}{logical. Should barplots with the aggregated
results be displayed on the screen?}

\item{export}{logical. Should detailed questions along with
individual results be exported to HTML files in a ZIP archive
for convenient import into OLAT?}
}
\value{
A \code{data.frame} with the detailed exam results is returned invisibly.
It is also written to a CSV file in the current directory, along with a ZIP
file containing the HTML reports (for upload into OLAT).
}
\description{
Evaluate OLAT exams produced with \code{\link[c403]{exams2olat}},
and carried out and exported in (Open)OLAT.
}
\details{
\code{olat_eval} is a companion function for \code{\link[c403]{exams2olat}}.
It evaluates the exams carried out in OLAT for further processing outside
of OLAT (in CSV format) and optionally exports detailed individual HTML
reports (in a ZIP archive) for reimport into OLAT.
}
\seealso{
\code{\link[c403]{exams2olat}}
}
\keyword{OLAT}
\keyword{utilities,}
