% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olat_eval.R
\name{olat_eval_adjust_lang}
\alias{olat_eval_adjust_lang}
\title{Adjust language of variables/columns}
\usage{
olat_eval_adjust_lang(x)
}
\arguments{
\item{x}{\code{data.frame} read from the xlsx file}
}
\value{
Returns the same \code{data.frame} (same dimension and data)
with adjusted names.
}
\description{
Modifies the names of the variables in the \code{dat.frame}
as read from the xlsx file (OpenOLAT). Converts the variable
names to English such that we do no longer have to care about
language in all other methods/functions.
}
\details{
Input 'x' is the data.frame as read from the xlsx file
which contains user meta information and the detailed information
about the individual questions of the test.
Problem: depending on the user language settings of OLAT the names
and order of the columns differs. This function takes input 'x' and
manipulates the variable or column names in a way that the rest
of the code (\code{\link{olat_eval}}) is not language dependent anymore.

\itemize{
\item{tries to guess the language by calling \code{\link{olat_eval_guess_lang}}}
\item{loads the search-replace-data.frame (internally)}
\item{search and replace variable names}
\item{return input object \code{x} with new varaible names}
}

Note: Even English to English will rename some of the variables.

The function uses the data set \code{olat_eval_lang} which is shipped
with the package (see \code{data("olat_eval_lang")}).
}
\author{
Reto
}
