% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olat_eval.R
\name{olat_eval_guess_lang}
\alias{olat_eval_guess_lang}
\title{Guess language of the OLAT results file}
\usage{
olat_eval_guess_lang(x)
}
\arguments{
\item{x}{character vector with column names}
}
\value{
Returns the language name (\code{"en"}, \code{"de"}) if
the function is able to guess the language. Else the script will stop.
}
\description{
Given a set of character vectors this function tries
to gess the language of the imported file. This allows to
evaluate exams from OpenOLAT in different languages.
Used by \code{\link[c403]{olat_eval}} to rename the columns
to make the evaluation independent from the language used when
exporting the results via OLAT.
}
\author{
Reto
}
