% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olat_feedback.R
\name{olat_feedback}
\alias{olat_feedback}
\alias{olat_feedback_render_one}
\title{Generate OLAT test feedback files (answer/solution)}
\usage{
olat_feedback(res, xexam, output = "olat_feedback.zip")

olat_feedback_render_one(res, xexam, i, show = FALSE,
  htmlfile = "Result.html")
}
\arguments{
\item{res}{data.frame, result from olat_eval}

\item{xexam}{list as returned from reading the rds file}

\item{output}{character, name of the zip archive file used to store the
individual feedback files}

\item{i}{integer, row index (which row in 'x' to render)}

\item{show}{logical, if set to \code{TRUE} the html file will
be opened in a browser}

\item{htmlfile}{character, name of the output file}
}
\value{
Returns the name of the temporary file created (html)
}
\description{
Similar to what \code{\link[c403]{olat_eval}} with the additional
flag \code{export = TRUE} does but in a different way. Shows ansers
and solutions. Creates a temporary html file.
}
\author{
Reto Stauffer
}
