% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olat_eval.R
\name{read_olat_results}
\alias{read_olat_results}
\title{Read OLAT Results from XLSX file}
\usage{
read_olat_results(file, xexam = NULL)
}
\arguments{
\item{file}{name of the file ...}

\item{xexam}{...}
}
\value{
...
}
\description{
Get OLAT test results (FIXME: eval support and cloze evaluation).
TODO: Write proper documentation.
}
