% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams2nops.R
\name{exams2nops}
\alias{exams2nops}
\title{Generation of NOPS Exams at Uni Innsbruck}
\usage{
exams2nops(
  file,
  n = 1L,
  dir = NULL,
  name = NULL,
  language = "de",
  title = "Klausur",
  course = "",
  institution = "Universit\\\\\\"at Innsbruck",
  logo = "uibk-logo-bw.png",
  date = Sys.Date(),
  replacement = TRUE,
  intro = NULL,
  blank = NULL,
  duplex = TRUE,
  pages = NULL,
  usepackage = NULL,
  encoding = "",
  startid = 1L,
  points = NULL,
  showpoints = FALSE,
  reglength = 8L,
  ...
)
}
\arguments{
\item{file}{character. A specification of a (list of) exercise files.}

\item{n}{integer. The number of copies to be compiled from \code{file}.}

\item{dir}{character. The default is either display on the screen or
the current working directory.}

\item{name}{character. A name prefix for resulting exercises and RDS file.}

\item{language}{character. Path to a DCF file with a language specification.
Currently, \code{"en"} and \code{"de"} are shipped with the \pkg{exams} package.}

\item{title}{character. Title of the exam, e.g., \code{"Statistische Datenanalyse"}.}

\item{course}{character. Course number, e.g., \code{"403001"}.}

\item{institution}{character. Name of the institution at which the exam is conducted.}

\item{logo}{character. Path to a logo image. If the logo is not found, it is simply omitted.}

\item{date}{character or \code{"Date"} object specifying the date of the exam.}

\item{replacement}{logical. Should a replacement exam sheet be included?}

\item{intro}{character with LaTeX code for introduction text at the beginning of the exam.}

\item{blank}{integer. Number of blank pages to be added at the end.
(Default is chosen to be half of the number of exercises.)}

\item{duplex}{logical. Should blank pages be added after the title page
(for duplex printing)?}

\item{pages}{character. Path(s) to additional PDF pages to be included
at the end of the exam.}

\item{usepackage}{character. Names of additional LaTeX packages to be included.}

\item{encoding}{character, passed to \code{\link[utils]{Sweave}}.}

\item{startid}{integer. Starting ID for the exam numbers (defaults to 1).}

\item{points}{integer. How many points should be assigned to each exercise? Note that this
argument overules any exercise points that are provided within the \code{expoints} tags
of the exercise files (if any). The vector of points supplied should either have
length 1 or the number of exercises in the exam.}

\item{showpoints}{logical. Should the PDF show the number of points associated with
each exercise (if specified in the Rnw/Rmd exercise or in \code{points})?}

\item{reglength}{integer. Number of digits in the registration ID. The
default is 8 and it can be increased up to 10.}

\item{...}{arguments passed on to \code{\link[exams]{exams2pdf}}.}
}
\value{
A list of exams as generated by \code{\link[exams]{xexams}} is returned
invisibly.
}
\description{
Interfaces \code{\link[exams]{exams2nops}} with different default
values as used at Uni Innsbruck.
}
\details{
\code{exams2nops} is a convenience interface for
\code{\link[exams]{exams2nops}} with somewhat different defaults:
German titles/descriptions, Uni Innsbruck logo, replacement sheets enabled,
duplex printing enabled.
}
\examples{
## load package and enforce par(ask = FALSE)
library("exams")
options(device.ask.default = FALSE)

## define an exams (= list of exercises)
myexam <- list(
  "boxplots",
  c("tstat", "ttest", "confint"),
  c("regression", "anova"),
  c("scatterplot", "boxhist"),
  "relfreq"
)

if(interactive()) {
  ## compile a single random exam (displayed on screen)
  exams2nops(c("tstat2", "anova", "boxplots"))
}

}
\seealso{
nops_eval nops_register
}
\keyword{utilities}
