\name{exams2nops}
\alias{exams2nops}
\alias{exams2olat}
\alias{exams2quiz}

\title{Exams Interfaces}

\description{
  Interfaces for generating exams at Uni Innsbruck.
}

\usage{
  exams2nops(file, n = 1L, dir = NULL, name = NULL,
    language = "de", title = "Klausur", course = "",
    institution = "Universit\\\\\\"at Innsbruck", logo = "uibk-logo-bw.png",
    date = Sys.Date(), replacement = TRUE, intro = NULL, blank = NULL,
    duplex = TRUE, pages = NULL, usepackage = NULL, encoding = "", startid = 1L,
    points = NULL, showpoints = FALSE, reglength = 8L, ...)

  exams2olat(file, n = 1L, dir = ".", name = "olattest", maxattempts = 1,
    cutvalue = 1000, solutionswitch = FALSE, 
    stitle = "Aufgabe", ititle = "Frage",
    adescription = "Bitte bearbeiten Sie folgende Aufgaben.",
    sdescription = "Bitte beantworten Sie folgende Frage.",
    eval = list(partial = FALSE, negative = FALSE),    
    \dots)

  exams2quiz(file, n = 1L, dir = ".", name = "VU", solution = FALSE,
    quiet = TRUE, encoding = "", \dots)
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}.}
  \item{dir}{character. The default is either display on the screen or
    the current working directory.}
  \item{name}{character. A name prefix for resulting exercises and RDS file.}
  \item{language}{character. Path to a DCF file with a language specification.
    Currently, \code{"en"} and \code{"de"} are shipped with the \pkg{exams} package.}
  \item{title}{character. Title of the exam, e.g., \code{"Statistische Datenanalyse"}.}
  \item{course}{character. Course number, e.g., \code{"403001"}.}
  \item{institution}{character. Name of the institution at which the exam is conducted.}
  \item{logo}{character. Path to a logo image. If the logo is not found, it is
    simply omitted.}
  \item{date}{character or \code{"Date"} object specifying the date of the exam.}
  \item{replacement}{logical. Should a replacement exam sheet be included?}
  \item{intro}{character with LaTeX code for introduction text at the beginning of the exam.}
  \item{blank}{integer. Number of blank pages to be added at the end.
    (Default is chosen to be half of the number of exercises.)}
  \item{duplex}{logical. Should blank pages be added after the title page
    (for duplex printing)?}
  \item{pages}{character. Path(s) to additional PDF pages to be included
    at the end of the exam.}
  \item{usepackage}{character. Names of additional LaTeX packages to be
    included.}
  \item{encoding}{character, passed to \code{\link[utils]{Sweave}}.}
  \item{startid}{integer. Starting ID for the exam numbers (defaults to 1).}
  \item{points}{integer. How many points should be assigned to each exercise? Note that this
    argument overules any exercise points that are provided within the \code{expoints} tags
    of the exercise files (if any). The vector of points supplied should either have
    length 1 or the number of exercises in the exam.}
  \item{showpoints}{logical. Should the PDF show the number of points associated with
    each exercise (if specified in the Rnw/Rmd exercise or in \code{points})?}
  \item{reglength}{integer. Number of digits in the registration ID. The
    default is 8 and it can be increased up to 10.}
  \item{maxattempts}{integer. The maximum attempts for one question, may also be set to
    \code{Inf}.}
  \item{cutvalue}{numeric. The cutvalue at which the exam is passed.}
  \item{solutionswitch}{logical. Should the question/item solutionswitch be enabled?}
  \item{stitle, ititle, adescription, sdescription, eval}{character. Descriptions
    for various titles/descriptions. Defaults to generic German titles.}
  \item{\dots}{arguments passed on to \code{\link[exams]{exams2pdf}} and
    \code{\link[exams]{exams2qti12}}, respectively.}
  \item{solution}{logical. Should the full solution be included?}
  \item{quiet}{logical. Should output be suppressed when calling
    \code{\link[utils]{Sweave}} and \code{\link[tools]{texi2dvi}}.}
}

\details{
  \code{exams2nops} is a convenience interface for \code{\link[exams]{exams2nops}}
  with UIBK-specific defaults.
  
  \code{exams2olat} is a convenience interface for \code{\link[exams]{exams2qti12}}
  with some tweaks for OLAT and German titles/descriptions.
  
  Both produce an RDS file as a by-product containing the \code{\link[exams]{xexams}}
  list.
  
  \code{exams2quiz} provides fallback support for quizzes in lecture sessions.
}

\value{
  A list of exams as generated by \code{\link[exams]{xexams}} is
  returned invisibly.
}

\examples{
## load package and enforce par(ask = FALSE)
library("exams")
options(device.ask.default = FALSE)

## define an exams (= list of exercises)
myexam <- list(
  "boxplots",
  c("tstat", "ttest", "confint"),
  c("regression", "anova"),
  c("scatterplot", "boxhist"),
  "relfreq"
)

## output directory
mydir <- tempdir()

## generate .zip with German OLAT exam in temporary directory
## using a few customization options
exams2olat(myexam, n = 3, dir = mydir,
  maxattempts = 3,
  num = list(digits = 1),
  mchoice = list(shuffle = TRUE, enumerate = FALSE)
)
dir(mydir)

if(interactive()) {
## compile a single random exam (displayed on screen)
exams2nops(c("tstat2", "anova", "boxplots"))

## compile a single random quiz (displayed on screen)
exams2quiz(rep("tstat2", 2))
}
}

\keyword{utilities}
