% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams2openolat.R
\name{exams2openolat}
\alias{exams2openolat}
\title{Exams to OpenOlat}
\usage{
exams2openolat(file, n = 1L, dir = ".", name = "olattest",
  maxattempts = 1, cutvalue = 1000, solutionswitch = FALSE,
  qti = "2.1", stitle = "Aufgabe", ititle = "Frage",
  adescription = "Bitte bearbeiten Sie folgende Aufgaben.",
  sdescription = "Bitte beantworten Sie folgende Frage.",
  eval = list(partial = FALSE, negative = FALSE), ...)
}
\description{
Interfaces \code{\link[exams]{exams2openolat}} with
slightly different default values.
TODO: Finish docstring
}
