% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nops_eval.R, R/nops_register.R
\name{nops_eval}
\alias{nops_eval}
\alias{nops_register}
\title{Evaluate NOPS Exams}
\usage{
nops_eval(
  register = Sys.glob("*.csv"),
  solutions = Sys.glob("*.rds"),
  scans = Sys.glob("nops_scan_*.zip"),
  points = NULL,
  eval = exams_eval(partial = FALSE, negative = 0.25),
  mark = c(0.5, 0.6, 0.75, 0.85),
  dir = ".",
  results = "nops_eval",
  html = NULL,
  col = hcl(c(0, 0, 60, 120), c(70, 0, 70, 70), 90),
  encoding = "latin1",
  language = "de",
  module = NULL,
  interactive = TRUE,
  string_scans = Sys.glob("nops_string_scan_*.zip"),
  string_points = seq(0, 1, 0.25)
)

nops_register(file = Sys.glob("*.xls*"), startid = 1L)
}
\arguments{
\item{register}{character. File name of a CSV file (semicolon-separated)
of the registered students, e.g., as produced by \code{nops_register}
based on the VIS registration lists. Must contain columns
\code{"Matrikelnr"}, \code{"Name"}, \code{"Account"}
(and \code{"LV"}, in case \code{module} marks should be computed).
The file name should not contain spaces, umlaut or other special characters
(something like \code{"GP-2014-02.csv"} is recommended).}

\item{solutions}{character. File name of the RDS exercise file
produced by \code{\link[c403]{exams2nops}}.}

\item{scans}{character. File name of the ZIP file with scanning results
(containing Daten.txt and PNG files) as produced by \code{\link[exams]{nops_scan}}
(or the FSS).}

\item{points}{numeric. Vector of points per exercise. By default read from
\code{solutions}.}

\item{eval}{list specification of evaluation policy as computed by
\code{\link[exams]{exams_eval}}.}

\item{mark}{logical or numeric. If \code{mark = FALSE}, no marks are computed.
Otherwise \code{mark} needs to be a numeric vector with four threshold values to
compute marks from 5 to 1. The thresholds can either be relative (all lower than 1)
or absolute. In case results exactly matching a threshold, the better mark is used.}

\item{dir}{character. File path to the output directory (the default being the
current working directory).}

\item{results}{character. Prefix for output files.}

\item{html}{character. File name for individual HTML files, by default
the same as \code{register} with suffix \code{.html}.}

\item{col}{character. Hex color codes used for exercises with
negative, neutral, positive, full solution.}

\item{encoding}{character. Encoding of \code{register}, typically either
\code{"latin1"} (default) or \code{"UTF-8"}.}

\item{language}{character. Path to a DCF file with a language specification.
Currently, \code{"en"} and \code{"de"} are shipped with the package.}

\item{module}{logical or numeric. Should module marks (in addition to the
exam marks) be computed? If this is numeric, this can be a vector of
two ECTS weights for the written exam and seminar, respectively (by
default equal weights of 0.5 and 0.5 are used). If \code{module} is not \code{FALSE},
\code{register} needs to contain a column \code{"LV"} with the seminar marks.}

\item{interactive}{logical. Should possible errors in the Daten.txt file
by corrected interactively? Requires the \pkg{png} package for full
interactivity.}

\item{string_scans}{character. Optional file name of the ZIP file with scanning results
of string exercise sheets (if any) containing Daten2.txt and PNG files as produced
by \code{\link[exams]{nops_scan}}.}

\item{string_points}{numeric. Vector of length 5 with points assigned
to string results.}

\item{file}{character, by default globbing is used. Can be a file name}

\item{startid}{integer, default \code{1L}}
}
\value{
A \code{data.frame} with the detailed exam results is returned
invisibly.  It is also written to a CSV file in the current directory, along
with a ZIP file containing the HTML reports (for upload into OLAT), and an XLSX
file (for importing the marks into VIS).
}
\description{
Evaluate NOPS exams produced with \code{\link[c403]{exams2nops}},
and scanned by \code{\link[exams]{nops_scan}}.
}
\details{
\code{nops_eval} is a companion function for
\code{\link[c403]{exams2nops}} and \code{\link[exams]{nops_scan}}.
It calls \code{\link[exams]{nops_eval}} from the \pkg{exams} package which evaluates
the scanned exams by computing the sums of the points achived and (if desired)
mapping them to marks (and to module marks). Furthermore HTML reports for each
individual student are generated for upload into OpenOlat. In addition to this
function from the \pkg{exams} package, the function adds the marks in the Uni
Innsbruck-specifc format in an Excel spreadsheet.

\code{nops_register} is another companion function for preprocessing the
registration lists that are provided by VIS. The function assigns random seats
for every student and saves the result in both CSV and XLSX format as well as a
tab-separated text file with the seat numbers for import into OLAT.  The
underlying workhorse function is \code{\link[c403]{read_vis}}.
}
\seealso{
\code{\link[c403]{exams2nops}}, \code{\link[exams]{nops_scan}}, \code{\link[exams]{nops_eval}}, \code{\link[c403]{read_vis}}
}
\keyword{utilities}
