\name{olat_eval}
\alias{olat_eval}

\title{Evaluate OLAT Exams}

\description{
  Evaluate OLAT exams produced with \code{\link[c403]{exams2olat}},
  and carried out and exported in (Open)OLAT.
}

\usage{
  olat_eval(file, plot = TRUE, export = FALSE)
}

\arguments{
  \item{file}{character. Base file name of RDS and XLS file with
    exam generated by \code{\link[c403]{exams2olat}} and exported
    OLAT results, respectively.}
  \item{plot}{logical. Should barplots with the aggregated
    results be displayed on the screen?}
  \item{export}{logical. Should detailed questions along with
    individual results be exported to HTML files in a ZIP archive
    for convenient import into OLAT?}
}

\details{
  \code{olat_eval} is a companion function for \code{\link[c403]{exams2olat}}.
  It evaluates the exams carried out in OLAT for further processing outside
  of OLAT (in CSV format) and optionally exports detailed individual HTML
  reports (in a ZIP archive) for reimport into OLAT.
}

\value{
  A \code{data.frame} with the detailed exam results is returned invisibly.
  It is also written to a CSV file in the current directory, along with a ZIP
  file containing the HTML reports (for upload into OLAT).
}

\seealso{\code{\link[c403]{exams2olat}}}

\keyword{utilities}
