% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olat_feedback.R
\name{olat_feedback}
\alias{olat_feedback}
\alias{olat_feedback_render_one}
\title{Generate OLAT test feedback files (answer/solution)}
\usage{
olat_feedback(res, xexam, name = "olat_feedback")

olat_feedback_render_one(res, xexam, i, htmlfile = "Result.html",
  show = FALSE)
}
\arguments{
\item{res}{data.frame, result from olat_eval}

\item{xexam}{list as returned from reading the rds file}

\item{name}{character, name of the test, will be used to name
the zip archive file and the html files}

\item{i}{integer, row index (which row in 'x' to render)}

\item{htmlfile}{character, name of the output file}

\item{show}{logical, if set to \code{TRUE} the html file will
be opened in a browser}
}
\value{
Returns the name of the zip file created.
}
\description{
Similar to what \code{\link[c403]{olat_eval}} does but in a more
detailed way. Creates html files for each participant with feedback
about his/her test results (including questions and solutions).
For OLAT tests use \code{\link[c403]{olat_feedback}},
for nops tests (written tests) use \code{\link[c403]{nops_feedback}}.
}
\author{
Reto Stauffer
}
