% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vis.R
\name{read_vis}
\alias{read_vis}
\alias{vis_register}
\title{Reading VIS Registrations}
\usage{
read_vis(file, ...)
}
\arguments{
\item{file}{character with file name of an XLS file from VIS.}

\item{...}{subset logical. Should students without confirmed registration be omitted?}
}
\value{
A \code{data.frame} with an additional attribute \code{"info"} providing
details about the type of course (\code{"LV"}) or exam (\code{"GP"}).
}
\description{
Read registration lists (for exams or courses) from the Excel export
of VIS (which actually may or may not be XLS or HTML files).
}
\details{
VIS offers Excel exports but in case of registration lists these are
typically HTML files containing an HTML table. These are read using the XML
package. However, some exports are actually Excel files which are read using
the xlsx package. In either case some basic cleaning is done and additional
meta-information is extracted.
}
\keyword{utilities}
