% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vis.R
\name{read_vis}
\alias{read_vis}
\alias{vis_register}
\title{Reading VIS Registrations}
\usage{
read_vis(file, ...)

vis_register(file = Sys.glob("*.xls"), subset = TRUE)
}
\arguments{
\item{file}{character with file name of an XLS file from VIS.}

\item{...}{additional arguments passed to \code{\link[xlsx]{read.xlsx}}.}

\item{subset}{logical. Should students without confirmed registration be omitted?}
}
\value{
A \code{data.frame} with an additional attribute \code{"info"} providing
details about the type of course (\code{"LV"}) or exam (\code{"GP"}).
}
\description{
Read registration lists (for exams or courses) from the Excel export
of VIS (which actually may or may not be XLS or HTML files).
}
\details{
VIS offers Excel exports but in case of registration lists these are
actually HTML files containing an HTML table. (Note that as of 2021 VIS offers an additional
\dQuote{real Excel} export.) HTMLtables are read using the XML
package. However, some exports are also converted to actual Excel files which are read using
the xlsx package. In either case some basic cleaning is done and additional
meta-information is extracted.

The \code{vis_register} function loops over reading several VIS exports and then
consolidates the resulting data frames.
}
\keyword{utilities}
