% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R, R/nops_eval.R
\name{mchoice2text}
\alias{mchoice2text}
\alias{uibkmark}
\title{Auxiliary Formatting Functions}
\usage{
mchoice2text(x, true = "\\\\\\\\textbf{Richtig}",
  false = "\\\\\\\\textbf{Falsch}")

uibkmark(x, factor = TRUE)
}
\arguments{
\item{x}{numeric (\code{uibkmark}) or logical (\code{mchoice2text}) vector.}

\item{true}{character, text for true results}

\item{false}{character, text for false results}

\item{factor}{logical. Should the result be a factor or a character?}
}
\description{
Auxiliary functions for formatting elements of exams.
\code{mchoice2text(x, true = "\\textbf{Richtig}", false = "\\textbf{Falsch}")}
}
\details{
The function \code{uibkmark} maps the numbers 1 to 5 to the mark labels
SGT1, GUT2, etc. as used by UIBK.

The function \code{mchoice2text} masks the exams function of the same name
in order to show German text.
}
\examples{
uibkmark(1:5)
mchoice2text(c(TRUE, FALSE))

}
\keyword{utilities}
